requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
coloredlogs==15.0.1
ddddocr==1.4.7
decorator==5.1.1
exceptiongroup==1.0.0rc9
flatbuffers==22.9.24
h11==0.14.0
humanfriendly==10.0
idna==3.4
mpmath==1.2.1
numpy==1.23.4
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
protobuf==4.21.8
pyee==8.1.0
pyparsing==3.0.9
PySocks==1.7.1
requests==2.28.1
selenium==4.5.0
self==2020.12.3
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.11.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
